pyproject.toml
[tool.poetry]
name = "PyEEPROM"
version = "0.1.0"
description = ""
authors = ["Daniel Montoya <da03@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
PySimpleGUI = "^4.60.3"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
